+++ app/services/project_request_statistics_service.py
@@ -30,7 +30,7 @@ class ProjectRequestStatisticsService:
     
     async def create_project_request_statistics(self, project_id: str, total_requests: Optional[int] = None,
                             completed_requests: Optional[int] = None, pending_requests: Optional[int] = None,
-                            failed_requests: Optional[int] = None, average_processing_time: Optional[float] = None,
+                            failed_requests: Optional[int] = None, average_processing_time: Optional[int] = None,
                             last_activity_at: Optional[str] = None, statistics_metadata: Optional[Dict[str, Any]] = None) -> ProjectRequestStatisticsModel:
         """Create a new project request statistics"""
         try:
